Entity clipboard representation rewritten.
The clipboard representations tries to come out with an HTML representation
that not only works but also seems natural to read, so it tries to avoid
any redundant data, removing those values coming from primitive attributes,
mixings, primitive defaults, A-Frame defaults and schema defaults.

To do that, it comes with a representation of the implicit value of each
component. I.e., the value coming from the sources mentioned above. Then it
gets the actual value of the component and computes minimal set of updates
needed to go from the implicit value to the actual value.