Telling inherited components apart form entity-owned components.
The optimizeComponents() routine had a path in which the component was removed
if the optimal update was `null`, pointing that there were no difference between
the implicit value and the current value. But this assumption made components
with only default attributes (as in <a-entity geometry=""></a-entity>) to be
removed.

The modification make getImpliciValue() to indicate if the component is
inherited from injected components or mixins, or if it is genuinely owned by the
entity. Just in the case the component is inherited and it does not need an
update, the component can be safely omitted.